Add method to generate multi-column IN SQL clauses
We can't use the `=ANY(ARRAY[...])` form for PostgreSQL, becaues it
can't resolve the comparison operator to use when the array contains
strings, as they have type "unknown".

PostgreSQL allows `(a,b) IN ((?, ?),(?, ?),...)`, but SQLite requires
the contents to be an actual subquery when there's more than one
tuple, so use `(a,b) IN (VALUES ...)` which works on both.